Fixed print-menu issue
Fixed the issue where the print menu would only show up for one second and could not be triggered to return.

Changed the logo to use both the logo from the Header and the navigation menu icon from the DashboatdViewPage.
The click event was moved to the navigation menu icon instead of the logo link.